+++ src/DependencyInjection/JobExtension.php
@@ -13,7 +13,7 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
-class ContaoHeimrichHannotJobExtension extends Extension
+class JobExtension extends Extension
 {
     public function getAlias()
     {
@@ -28,5 +28,6 @@ class ContaoHeimrichHannotJobExtension extends Extension
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
 
         $loader->load('listener.yml');
+        $loader->load('services.yml');
     }
 }