+++ src/Item/JobItemTrait.php
@@ -2,6 +2,8 @@
 
 namespace HeimrichHannot\JobBundle\Item;
 
+use Contao\ContentModel;
+use Contao\Controller;
 use Contao\StringUtil;
 use Contao\System;
 
@@ -16,8 +18,7 @@ trait JobItemTrait
     {
         $instances = [];
 
-        if (null !== ($archive  = System::getContainer()->get('huh.utils.model')->findModelInstanceByPk('tl_job_archive', $this->pid)))
-        {
+        if (null !== ($archive = System::getContainer()->get('huh.utils.model')->findModelInstanceByPk('tl_job_archive', $this->pid))) {
             $instances[] = $archive;
         }
 
@@ -27,23 +28,96 @@ trait JobItemTrait
             'memberContacts', $instances
         );
 
-        if (!$contacts)
-        {
+        if (!$contacts) {
             return null;
         }
 
         $ids = StringUtil::deserialize($contacts, true);
 
-        if (empty($ids))
-        {
+        if (empty($ids)) {
             return null;
         }
 
-        if (null === ($members = System::getContainer()->get('huh.utils.model')->findMultipleModelInstancesByIds('tl_member', $ids)))
-        {
+        if (null === ($members = System::getContainer()->get('huh.utils.model')->findMultipleModelInstancesByIds('tl_member', $ids))) {
             return null;
         }
 
         return $members->fetchAll();
+    }
+
+    /**
+     * Get all enclosures.
+     *
+     * @return array|null
+     */
+    public function getEnclosures(): ?array
+    {
+        if (true === $this->enclosure) {
+            return null;
+        }
+
+        $template = new \stdClass();
+        Controller::addEnclosuresToTemplate($template, $this->getRaw());
+
+        return $template->enclosure;
+    }
+
+    /**
+     * Compile the job content (tl_content).
+     *
+     * @return string
+     */
+    public function getContent(): string
+    {
+        $strText = '';
+
+        /**
+         * @var ContentModel
+         */
+        $adapter = $this->getManager()->getFramework()->getAdapter(ContentModel::class);
+
+        if (null !== ($elements = $adapter->findPublishedByPidAndTable($this->id, $this->getDataContainer()))) {
+            foreach ($elements as $element) {
+                try {
+                    $strText .= Controller::getContentElement($element->id);
+                } catch (\ErrorException $e) {
+                }
+            }
+        }
+
+        return $strText;
+    }
+
+    /**
+     * Check if the job has content (tl_content).
+     *
+     * @return bool
+     */
+    public function hasContent(): bool
+    {
+        /** @var ContentModel $adapter */
+        $adapter = $this->getManager()->getFramework()->getAdapter(ContentModel::class);
+
+        return $adapter->countPublishedByPidAndTable($this->id, $this->getDataContainer());
+    }
+
+    /**
+     * Check if the job has teaser text.
+     *
+     * @return bool
+     */
+    public function hasTeaser(): bool
+    {
+        return '' !== $this->teaser;
+    }
+
+    /**
+     * Compile the job text.
+     *
+     * @return string
+     */
+    public function getTeaser(): string
+    {
+        return StringUtil::encodeEmail(StringUtil::toHtml5($this->teaser));
     }
 }