+++ src/Resources/contao/dca/tl_job_archive.php
@@ -6,7 +6,7 @@ $GLOBALS['TL_DCA']['tl_job_archive'] = [
         'switchToEdit'      => true,
         'enableVersioning'  => true,
         'onload_callback'   => [
-            ['huh.job.listener.job_archive_callback', 'checkPermission'],
+            ['huh.job.data_container.job_archive', 'checkPermission'],
         ],
         'onsubmit_callback' => [
             ['huh.utils.dca', 'setDateAdded'],
@@ -49,20 +49,20 @@ $GLOBALS['TL_DCA']['tl_job_archive'] = [
                 'label'           => &$GLOBALS['TL_LANG']['tl_job_archive']['editheader'],
                 'href'            => 'act=edit',
                 'icon'            => 'header.gif',
-                'button_callback' => ['huh.job.listener.job_archive_callback', 'editHeader']
+                'button_callback' => ['huh.job.data_container.job_archive', 'editHeader']
             ],
             'copy'       => [
                 'label'           => &$GLOBALS['TL_LANG']['tl_job_archive']['copy'],
                 'href'            => 'act=copy',
                 'icon'            => 'copy.gif',
-                'button_callback' => ['huh.job.listener.job_archive_callback', 'copyArchive']
+                'button_callback' => ['huh.job.data_container.job_archive', 'copyArchive']
             ],
             'delete'     => [
                 'label'           => &$GLOBALS['TL_LANG']['tl_job_archive']['copy'],
                 'href'            => 'act=delete',
                 'icon'            => 'delete.gif',
                 'attributes'      => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\'))return false;Backend.getScrollOffset()"',
-                'button_callback' => ['huh.job.listener.job_archive_callback', 'deleteArchive']
+                'button_callback' => ['huh.job.data_container.job_archive', 'deleteArchive']
             ],
             'show'       => [
                 'label' => &$GLOBALS['TL_LANG']['tl_job_archive']['show'],